backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
torch
torchvision
torchaudio
pillow
transformers
faiss-cpu
git+https://github.com/SauravMaheshkar/samv2.git
redis
timm
einops
pytest